Fix unconvert lint error in device_unix.go
Remove unnecessary uint64 conversion for stat.Dev.

+++ internal/scanner/device_unix.go
@@ -13,5 +13,5 @@ func getDeviceID(info os.FileInfo) (uint64, bool) {
 	if !ok {
 		return 0, false
 	}
-	return uint64(stat.Dev), true
+	return stat.Dev, true
 }
